git-svn-id: http://svn.code.sf.net/p/fossology/code/branches/new-build-system@945 a5643f74-2041-0410-a2b9-94ffef304a07

+++ ui/plugins/agent-license-once-compare.php
@@ -24,24 +24,22 @@
 global $GlobalReady;
 if (!isset($GlobalReady)) { exit; }
 
-class agent_license_once extends FO_Plugin
+class agent_license_once_compare extends FO_Plugin
 {
-  public $Name       = "agent_license_once";
-  public $Title      = "One-Shot License Analysis";
+  public $Name       = "agent_license_once_compare";
+  public $Title      = "One-Shot License Comparison";
   public $Version    = "1.0";
-  public $Dependency = array("view","view-license");
+  public $Dependency = array("db","view","view-license");
   public $NoHTML     = 0;
-  /** For anyone to access, without login, use: **/
-  // public $DBaccess   = PLUGIN_DB_NONE;
-  // public $LoginFlag  = 0;
   /** To require login access, use: **/
   public $DBaccess   = PLUGIN_DB_ANALYZE;
   public $LoginFlag  = 1;
 
   /*********************************************
    AnalyzeOne(): Analyze one uploaded file.
+   Returns 0 on success, !0 on failure.
    *********************************************/
-  function AnalyzeOne ($Highlight)
+  function AnalyzeOne ($Highlight,$LicList)
   {
     global $Plugins;
     global $AGENTDIR;
@@ -53,6 +51,7 @@ class agent_license_once extends FO_Plugin
     /* move the temp file */
     $TempFile = $_FILES['licfile']['tmp_name'];
     $TempCache = $TempFile . ".bsam";
+    $TempLics = $TempFile . ".lic.bsam";
     // print "TempFile=$TempFile   TempCache=$TempCache\n";
 
     /* Create cache file */
@@ -60,8 +59,36 @@ class agent_license_once extends FO_Plugin
     system($Sys);
     // print "Cached file $TempCache = " . filesize($TempCache) . " bytes.\n";
 
+    /* Create cache file for licenses */
+    global $DB;
+    $SQL = "";
+    for($i=0; !empty($LicList[$i]); $i++)
+      {
+      if (!empty($SQL)) { $SQL .= " OR"; }
+      $SQL .= " lic_id = '" . intval($LicList[$i]) . "'";
+      }
+    if (empty($SQL))
+	{
+	print "No comparison licenses found.";
+	return(1);
+	}
+    $SQL = "SELECT DISTINCT lic_name FROM agent_lic_raw WHERE $SQL;";
+    $Lics = $DB->Action($SQL);
+    if (empty($Lics))
+	{
+	print "No comparison licenses found.";
+	return(1);
+	}
+    chdir("$DATADIR/agents/licenses/");
+    for($i=0; !empty($Lics[$i]['lic_name']); $i++)
+	{
+	$Filename = $Lics[$i]['lic_name'];
+	$Sys = "$AGENTDIR/Filter_License -O '$Filename' >> '$TempLics'";
+	system($Sys);
+	}
+
     /* Create bsam results */
-    $Sys = "$AGENTDIR/bsam-engine -L 20 -A 0 -B 60 -G 15 -M 10 -E -O t '$TempCache' '$DATADIR/agents/License.bsam'";
+    $Sys = "$AGENTDIR/bsam-engine -L 20 -A 0 -B 60 -G 15 -M 10 -E -O t '$TempCache' '$TempLics'";
     $Fin = popen($Sys,"r");
     $LicSummary = array();
     $LicNum = -1;
@@ -151,6 +178,7 @@ class agent_license_once extends FO_Plugin
 	} /* while read a line */
       } /* while read from bsam */
     pclose($Fin);
+    unlink($TempLics);
 
     if ($Highlight)
       {
@@ -226,17 +254,13 @@ class agent_license_once extends FO_Plugin
     /* Only register with the menu system if the user is logged in. */
     if (!empty($_SESSION['User']))
       {
-      if (@$_SESSION['UserLevel'] >= PLUGIN_DB_ANALYZE)	// Debugging changes to license analysis
-	{
-	menu_insert("Main::Upload::One-Shot License",$this->MenuOrder,$this->Name,$this->MenuTarget);
-	}
       if (@$_SESSION['UserLevel'] >= PLUGIN_DB_DEBUG)	// Debugging changes to license analysis
 	{
-	$URI = $this->Name . Traceback_parm_keep(array("format","pfile"));
+	$URI = $this->Name . Traceback_parm_keep(array("format","pfile","item","ufile"));
 	menu_insert("View::[BREAK]",100);
-	menu_insert("View::One-Shot",101,$URI,"One-shot, real-time license analysis");
+	menu_insert("View::Recompare",101,$URI,"One-shot, real-time license recomparison");
 	menu_insert("View-Meta::[BREAK]",100);
-	menu_insert("View-Meta::One-Shot",101,$URI,"One-shot, real-time license analysis");
+	menu_insert("View-Meta::Recompare",101,$URI,"One-shot, real-time license recomparison");
 	}
       }
   } // RegisterMenus()
@@ -255,28 +279,30 @@ class agent_license_once extends FO_Plugin
 	break;
       case "HTML":
 	/* If this is a POST, then process the request. */
-	$Highlight = GetParm('highlight',PARM_INTEGER); // may be null
 	/* You can also specify the file by pfile_pk */
 	$PfilePk = GetParm('pfile',PARM_INTEGER); // may be null
-	if (file_exists(@$_FILES['licfile']['tmp_name']))
+	$UfilePk = GetParm('ufile',PARM_INTEGER); // may be null
+	$Item = GetParm('item',PARM_INTEGER); // may be null
+	$LicList = GetParm('liclist',PARM_RAW); // may be null
+	$Highlight=1; /* Always highlight. */
+	if (is_array($LicList) && !empty($LicList[0]) && file_exists(@$_FILES['licfile']['tmp_name']))
 	  {
 	  if ($_FILES['licfile']['size'] <= 1024*1024*10)
 	    {
 	    /* Size is not too big.  */
-	    print $this->AnalyzeOne($Highlight) . "\n";
+	    print $this->AnalyzeOne($Highlight,$LicList) . "\n";
 	    }
 	  if (!empty($_FILES['licfile']['unlink_flag']))
 	    { unlink($_FILES['licfile']['tmp_name']); }
 	  return;
 	  }
-	else if (!empty($PfilePk) && !empty($DB))
+	else if (is_array($LicList) && !empty($LicList[0]) && !empty($PfilePk) && !empty($DB))
 	  {
 	  /* Get the pfile info */
 	  $Results = $DB->Action("SELECT * FROM pfile WHERE pfile_pk = '$PfilePk';");
 	  if (!empty($Results[0]['pfile_pk']))
 	    {
 	    global $LIBEXECDIR;
-	    $Highlight=1; /* processing a pfile? Always highlight. */
 	    $Repo = $Results[0]['pfile_sha1'] . "." . $Results[0]['pfile_md5'] . "." . $Results[0]['pfile_size'];
 	    $Repo = trim(shell_exec("$LIBEXECDIR/reppath files '$Repo'"));
 	    $_FILES['licfile']['tmp_name'] = $Repo;
@@ -284,7 +310,7 @@ class agent_license_once extends FO_Plugin
 	    if ($_FILES['licfile']['size'] <= 1024*1024*10)
 	      {
 	      /* Size is not too big.  */
-	      print $this->AnalyzeOne($Highlight) . "\n";
+	      print $this->AnalyzeOne($Highlight,$LicList) . "\n";
 	      }
 	    /* Do not unlink the or it will delete the repo file! */
 	    if (!empty($_FILES['licfile']['unlink_flag']))
@@ -294,24 +320,55 @@ class agent_license_once extends FO_Plugin
 	  }
 
 	/* Display instructions */
-	$V .= "This analyzer allows you to upload a single file for license analysis.\n";
+	$V .= "This analyzer allows you to upload a single file for license analysis and select the licenses to compare against.\n";
 	$V .= "The analysis is done in real-time.\n";
-	$V .= "The limitations:\n";
+	$V .= "<P>The limitations:\n";
 	$V .= "<ul>\n";
 	$V .= "<li>The analysis is done in real-time. Large files may take a while. This method is not recommended for files larger than a few hundred kilobytes.\n";
+	$V .= "<li>The analysis is done in real-time. Selecting many licenses to compare against can take a long time.\n";
 	$V .= "<li>Files that contain files are <b>not</b> unpacked. If you upload a 'zip' or 'deb' file, then the binary file will be scanned for licenses and nothing will likely be found.\n";
-	$V .= "<li>Results are <b>not</b> stored. As soon as you get your results, your uploaded file is removed from the system.\n";
+	$V .= "<li>Results are <b>not</b> stored. As soon as you get your results, your analysis is removed from the system.\n";
 	$V .= "</ul>\n";
 
-	/* Display the form */
 	$V .= "<form enctype='multipart/form-data' method='post'>\n";
 	$V .= "<ol>\n";
-	$V .= "<li>Select the file to upload:<br />\n";
-	$V .= "<input name='licfile' size='60' type='file' /><br />\n";
-	$V .= "<b>NOTE</b>: Files larger than 100K will be discarded and not analyzed.<P />\n";
-	$V .= "<li><input type='checkbox' name='highlight' value='1'>Check if you want to see the highlighted licenses.\n";
-	$V .= "Unchecked returns a simple list that summarizes the identified license types.";
-	$V .= "<P />\n";
+	/* Display the form */
+	if (empty($PfilePk))
+	  {
+	  $V .= "<li>Select the file to upload:<br />\n";
+	  $V .= "<input name='licfile' size='60' type='file' /><br />\n";
+	  $V .= "<b>NOTE</b>: Files larger than 100K will be discarded and not analyzed.<P />\n";
+	  $V .= "<input type='hidden' name='pfile' value='$PfilePk'>";
+	  $V .= "<input type='hidden' name='ufile' value='$UfilePk'>";
+	  $V .= "<input type='hidden' name='item' value='$Item'>";
+	  }
+	else
+	  {
+	  $V .= "<li>This is the selected file to re-analyze:<br />\n";
+	  $V .= Dir2Browse("license",$Item,$UfilePk) . "<P />\n";
+	  }
+
+	$V .= "<li>Select one or more licenses to compare against:<br>\n";
+	$V .= "<table border=0><tr>";
+	$V .= "<td><select size='10' multiple='multiple' id='liclist' name='liclist[]'>\n";
+	$Lics = $DB->Action("SELECT DISTINCT lic_name,lic_id FROM agent_lic_raw ORDER BY lic_name;");
+	for($i=0; !empty($Lics[$i]['lic_name']); $i++)
+	  {
+	  $V .= "<option value='" . $Lics[$i]['lic_id'] . "'>" . $Lics[$i]['lic_name'] . "</option>\n";
+	  }
+	$V .= "</select>\n";
+	$Uri = "if (document.getElementById('liclist').value) { window.open('";
+	$Uri .= Traceback_uri();
+	$Uri .= "?mod=view-license";
+	$Uri .= "&format=flow";
+	$Uri .= "&lic=";
+	$Uri .= "' + document.getElementById('liclist').value + '";
+	$Uri .= "&licset=";
+	$Uri .= "' + document.getElementById('liclist').value";
+	$Uri .= ",'License','width=600,height=400,toolbar=no,scrollbars=yes,resizable=yes'); }";
+	$V .= "</td><td>";
+	$V .= "<a href='#' onClick=\"$Uri\">View</a>\n";
+	$V .= "</td></tr></table>";
 
 	$V .= "</ol>\n";
 	$V .= "<input type='hidden' name='showheader' value='1'>";
@@ -328,6 +385,6 @@ class agent_license_once extends FO_Plugin
     return;
   }
 };
-$NewPlugin = new agent_license_once;
+$NewPlugin = new agent_license_once_compare;
 $NewPlugin->Initialize();
 ?>
